requirements.txt
firebase_admin
psycopg2_binary